% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendis.r
\name{gendis}
\alias{gendis}
\title{Calculates a generalized discriminant function}
\usage{
gendis(
  population = "population",
  sex = "sex",
  measurements = "other_variables",
  verbose = FALSE,
  data
)
}
\arguments{
\item{population}{a name of the variable for the populations in the data (default "population")}

\item{sex}{a name of the variable indicating the two classes to distinguish in the data (default "sex")
(0 vs 1 or "female" vs "male")}

\item{measurements}{character ("other_variables", default)  or character vector with names of
measurement variables. \code{gendis} maintains the order of the names.}

\item{verbose}{logical (default = FALSE)}

\item{data}{data frame with variables}
}
\value{
An object of class gendis which is a named list, among which,
\item{population}{name of variable indicating populations}
\item{sex}{name of variable indicating the two sexes or classes}
\item{classnames}{names for the classes of sex (level or value)}
\item{measurements}{names of the variables in the GDF}
\item{GDF}{the Generalized Discriminant Function, matrix with two columns differing in scaling of the GDF}
\item{mean.male}{overall mean of males (the second level of factor(sex))}
\item{mean.female}{overall mean of females (the first level of factor(sex))}
\item{within.sd}{overall within standard deviation}
\item{cov_overall}{overall within-group covariance matrix}
\item{means.male}{mean of males per population}
\item{means.female}{mean of females per population}
\item{within.sds}{within standard deviation per population}
\item{ind_mv}{number of males and females per population}
\item{cov_list}{within-group covariance matrix per population}
\item{Nind}{number of individuals}
\item{Np}{number of populations}
}
\description{
\code{gendis} calculates a generalized discriminant function to distinguish two classes,
typically sexes (male and female birds)
based on measurements of a number of indicators for individuals from each of the two sexes
from a series of different populations in which individuals may have a different mean size but
a common-within covariance matrix.
}
\examples{
data("fulmarin")
names(fulmarin)
result <- gendis(population = "population", sex = "sex",
                 measurements = "other_variables", verbose = FALSE ,  data=fulmarin )
result$GDF
summary(result)
print(result)

# populations may have names:
fulmarin$pop <- factor(c("a1","a2","a3","a4","a5","a6")[fulmarin$population])
levels(fulmarin$pop)
names(fulmarin)
result2 <- gendis(population = "pop", sex = "sex",
                  measurements = c("HB","B","TL","CL"), verbose = FALSE ,  data=fulmarin )
# all equal should not give numeric differences.
#all.equal(result, result2)

result2$GDF - result$GDF



}
\references{
van Franeker, J A. ter Braak, C J F. 1993.
A generalized discriminant for sexing fulmarine petrels from
external measurements. The Auk 110: pp 492-502, https://doi.org/10.2307/4088413 https://edepot.wur.nl/249350
}
\seealso{
\code{\link{predict.gendis}}, \code{\link{summary.gendis}}, \code{\link{print.gendis}}.
}
