% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JanMayenBirds.R
\docType{data}
\name{JanMayenBirds}
\alias{JanMayenBirds}
\title{Fulmarin petrels with unknown sex from Jan Mayen}
\description{
The data frame \code{JanMayenBirds} contains measurements on Fulmarine petrels birds from Jan Mayen.
 From the first 32 birds the sex is known by dissection, from the remaining 162 birds the sex is unknown.
\itemize{
\item\code{JAFCODE} bird code (character)
\item\code{LOCATION} location (character)
\item\code{DATE} measurement date (character)
\item\code{DISSEX} 0 is female; 1 is male
\item\code{HB} Head Length (mm)
\item\code{BD2} Bill Depth at gonys (mm)
\item\code{TL} Tarsus Length (mm)
\item\code{CL} Culmen Length (mm.
}
}
\references{
van Franeker, J A. ter Braak, C J F. 1993.
A generalized discriminant for sexing fulmarine petrels from
external measurements. The Auk 110: pp 492-502, https://doi.org/10.2307/4088413 https://edepot.wur.nl/249350
}
\author{
Jan Andries van Franeker (\email{jan.vanfraneker@wur.nl})
}
