% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gendis.r
\name{predict.gendis}
\alias{predict.gendis}
\title{Predict function using a generalized discriminant function}
\usage{
\method{predict}{gendis}(object, newdata, type = object$sex, verbose = FALSE, ...)
}
\arguments{
\item{object}{an object of class gendis, typically created with \code{\link{gendis}}}

\item{newdata}{a data frame with measurements on (new) individuals with variables used to create \code{object}.
The data should be from a single population. If your data are from multiple populations, use \code{predict} for each
subset.}

\item{type}{what to predict : the sex or class of each individual (default),
the generalized discriminant scores with cutpoint ("GDF" or "GDFscore") or
the full output of the unmixing algorithm  ("cutpoint")}

\item{verbose}{logical (default = FALSE). If TRUE a plot of the density of the GDF is produced.}

\item{...}{other optional arguments}
}
\value{
See argument \code{type}.
}
\description{
\code{predict.gendis} applies a generalized discriminant function created with \code{\link{gendis}}
 to predict the sex (class) of each individual with measurements in \code{newdata}. From the  \code{\link{gendis}}
 object, the coefficients that define the generalized discriminant function (GDF) are to applied to the
 \code{newdata} to obtain the discriminant scores.
}
\details{
The discriminant scores (a linear combination of the measurements of the \code{newdata}
using the GDF coefficients) are subjected to an unmixing algorithm. This algorithm (\code{\link{unmix}}) generates a cutpoint
below which individuals are predicted to be female (level 1 of \code{factor(sex)}) and above which they are predicted to be
male (level 2 of \code{factor(sex)}). The cutpoint is at the point of intersection of two normal densities with unequal
means and variances fitted to the discriminant scores (see \code{\link{unmix} for details}).
}
\examples{
data("fulmarin")
str(fulmarin)
result <- gendis(population = "population", sex = "sex",
                 measurements = "other_variables", verbose = FALSE ,  data=fulmarin )
data("JanMayenBirds")
sex.predict <- predict(result, newdata = JanMayenBirds, verbose = TRUE)
# one false prediction: (number 32)
data.frame(sex = JanMayenBirds$DISSEX,  sex.predict)[seq(from=2, to = 37, by =5),]

predict(result, JanMayenBirds )
# same as default above
predict(result, JanMayenBirds, type = result$sex, verbose = FALSE)
# GDF score with cutpoint
predict(result, JanMayenBirds, type = "GDF", verbose = FALSE)
# unmix results only
predict(result, JanMayenBirds, type = "cutpoint", verbose = TRUE)



}
\references{
van Franeker, J A. ter Braak, C J F. 1993.
A generalized discriminant for sexing fulmarine petrels from
external measurements. The Auk 110: pp 492-502, https://doi.org/10.2307/4088413 https://edepot.wur.nl/249350
}
