% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gendis.r
\name{summary.gendis}
\alias{summary.gendis}
\title{Summary of a generalized discriminant analysis}
\usage{
\method{summary}{gendis}(object, ...)
}
\arguments{
\item{object}{an object of class gendis, created by \code{\link{gendis}}.}

\item{...}{other optional arguments.}
}
\value{
GDF
}
\description{
\code{summary.gendis} summarizes the results of \code{\link{gendis}}.
}
\examples{
data("fulmarin")
names(fulmarin)
result <- gendis(population = "population", sex = "sex",
                 measurements = "other_variables", verbose = FALSE ,  data=fulmarin )
result$GDF
summary(result)
print(result)

# populations may have names:
fulmarin$pop <- factor(c("a1","a2","a3","a4","a5","a6")[fulmarin$population])
levels(fulmarin$pop)
names(fulmarin)
result2 <- gendis(population = "pop", sex = "sex",
                  measurements = c("HB","B","TL","CL"), verbose = FALSE ,  data=fulmarin )
# all equal should not give numeric differences.
#all.equal(result, result2)

result2$GDF - result$GDF



}
\references{
van Franeker, J A. ter Braak, C J F. 1993.
A generalized discriminant for sexing fulmarine petrels from
external measurements. The Auk 110: pp 492-502ter Braak (2019)
}
\seealso{
\code{\link{gendis}}, \code{\link{print.gendis}},\code{\link{predict.gendis}}.
}
